README.md
This repository contains my solution to the QR Code Component Challenge from Frontend Mentor. The purpose of this README is to explain the project setup, technologies used, and future improvements I plan to make. It provides insight into how I approached the challenge and the steps I took to create a responsive QR code component.